requirements.txt
loguru~=0.7.3
dynaconf~=3.2.7
pydantic~=2.10.6
fastapi[standard]~=0.115.12
gradio~=5.9.1
uvicorn~=0.34.0
typing-extensions~=4.12.2
python-dotenv>=1.1.0
numpy<=1.26.4
fastrtc
aiortc~=1.12.0
pillow~=11.1.0
onnxruntime~=1.20.1
funasr~=1.2.3
torch==2.8.0
torchaudio
openai>=1.72.0
modelscope>=1.25.0

# CosyVoice
librosa~=0.10.2
conformer==0.3.2
gdown==5.1.0
hyperpyyaml==1.2.2
inflect>=6.5.0
lightning==2.2.4
matplotlib==3.7.5
openai-whisper==20240930
pyarrow>=19.0.1
pynini==2.1.6
pyworld==0.3.4
wget==3.2
wetext